library/requirements.txt
asgiref==3.8.1
Django==4.2.16
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
PyJWT==2.9.0
sqlparse==0.5.1
tzdata==2024.2
gunicorn==22.0.0
psycopg2-binary==2.9.7
djangorestframework-gis==1.1
Pillow==9.4.0
requests==2.31.0
celery==5.4.0
django-celery-results==2.5.1